Cardano.Ledger.Pretty

Documentation

ppString :: String -> Doc a

ppDouble :: Double -> Doc a

ppInteger :: Integer -> Doc a

ppRational :: Rational -> Doc a

ppFloat :: Float -> Doc a

ppNatural :: Natural -> Doc a

ppWord64 :: Word64 -> Doc a

ppWord32 :: Word32 -> Doc a

ppWord8 :: Word8 -> Doc a

ppWord16 :: Word16 -> Doc a

ppFixedPoint :: FixedPoint -> Doc a

ppPair :: (t1 -> PDoc) -> (t2 -> PDoc) -> (t1, t2) -> PDoc

ppSignedDSIGN :: Show a => a -> PDoc

ppBool :: Bool -> Doc a

ppInt :: Int -> Doc a

isEmpty :: Doc ann -> Bool

putDoc :: Doc ann -> IO ()

newtype PrettyAnn

Constructors

Width Int

type Ann = [PrettyAnn]

type PDoc = Doc Ann

text :: Text -> Doc ann

long_bech32 :: ByteString -> Text

lazy_bech32 :: ByteString -> Text

ppLong :: ByteString -> PDoc

ppLazy :: ByteString -> PDoc

equate :: Doc a -> Doc a -> Doc a

x == y

arrow :: (Doc a, Doc a) -> Doc a

x -> y

ppSexp :: Text -> [PDoc] -> PDoc

ppSexp x [w,y,z] --> (x w y z)

ppSexp' :: PDoc -> [PDoc] -> PDoc

ppRecord :: Text -> [(Text, PDoc)] -> PDoc

ppRecord name [("a",x),("b",y),("c",z)] --> name { a = x, b = y, c = z }

ppRecord' :: PDoc -> [(Text, PDoc)] -> PDoc

puncLeft :: Doc ann -> [Doc ann] -> Doc ann -> Doc ann -> Doc ann

Vertical layout with commas aligned on the left hand side

ppSet :: (x -> Doc ann) -> Set x -> Doc ann

ppList :: (x -> Doc ann) -> [x] -> Doc ann

ppStrictSeq :: (a -> Doc ann) -> StrictSeq a -> Doc ann

ppStrictMaybe :: (x -> Doc ann) -> StrictMaybe x -> Doc ann

ppMaybe :: (x -> Doc ann) -> Maybe x -> Doc ann

ppAssocList :: PDoc -> (k -> PDoc) -> (v -> PDoc) -> [(k, v)] -> PDoc

ppMap' :: PDoc -> (k -> PDoc) -> (v -> PDoc) -> Map k v -> PDoc

ppMap :: (k -> PDoc) -> (v -> PDoc) -> Map k v -> PDoc

ppVMap :: (Vector kv k, Vector vv v) => (k -> PDoc) -> (v -> PDoc) -> VMap kv vv k v -> PDoc

class PrettyA t where

Methods

prettyA :: t -> PDoc

ppTrip :: Triple crypto -> PDoc

ppUnifiedMap :: UnifiedMap crypto -> PDoc

ppLastAppliedBlock :: LastAppliedBlock c -> PDoc

ppHashHeader :: HashHeader c -> PDoc

ppWithOrigin :: (t -> PDoc) -> WithOrigin t -> PDoc

ppBHBody :: Crypto c => BHBody c -> PDoc

ppPrevHash :: PrevHash c -> PDoc

ppBHeader :: Crypto c => BHeader c -> PDoc

ppBlock :: (PrettyA (TxSeq era), PrettyA h) => Block h era -> PDoc

ppPoolDistr :: PoolDistr c -> PDoc

ppIndividualPoolStake :: IndividualPoolStake c -> PDoc

ppRewardUpdate :: RewardUpdate crypto -> PDoc

ppRewardSnapShot :: RewardSnapShot crypto -> PDoc

ppPoolRewardInfo :: PoolRewardInfo crypto -> PDoc

ppFreeVars :: FreeVars crypto -> PDoc

ppAns :: RewardAns crypto -> PDoc

ppRewardPulser :: Pulser crypto -> PDoc

ppPulsingRewUpdate :: PulsingRewUpdate crypto -> PDoc

type CanPrettyPrintLedgerState era = (PrettyA (TxOut era), PrettyA (PParams era), PrettyA (State (EraRule "PPUP" era)))

Constraints needed to ensure that the ledger state can be pretty printed.

ppAccountState :: AccountState -> PDoc

ppDPState :: DPState crypto -> PDoc

ppDState :: DState crypto -> PDoc

ppFutureGenDeleg :: FutureGenDeleg crypto -> PDoc

ppInstantaneousRewards :: InstantaneousRewards crypto -> PDoc

ppPPUPState :: PrettyA (PParamsDelta era) => PPUPState era -> PDoc

ppPState :: PState crypto -> PDoc

ppRewardAccounts :: Map (Credential 'Staking crypto) Coin -> PDoc

ppRewardType :: RewardType -> PDoc

ppReward :: Reward crypto -> PDoc

ppLeaderOnlyReward :: LeaderOnlyReward crypto -> PDoc

ppIncrementalStake :: IncrementalStake crypto -> PDoc

ppUTxOState :: CanPrettyPrintLedgerState era => UTxOState era -> PDoc

ppEpochState :: CanPrettyPrintLedgerState era => EpochState era -> PDoc

ppNewEpochState :: CanPrettyPrintLedgerState era => NewEpochState era -> PDoc

ppLedgerState :: CanPrettyPrintLedgerState era => LedgerState era -> PDoc

ppPerformanceEstimate :: PerformanceEstimate -> PDoc

ppNonMyopic :: NonMyopic crypto -> PDoc

ppStakeShare :: StakeShare -> PDoc

ppHistogram :: Histogram -> PDoc

ppLogWeight :: LogWeight -> PDoc

ppLikelihood :: Likelihood -> PDoc

ppStake :: Stake crypto -> PDoc

ppBlocksMade :: BlocksMade crypto -> PDoc

ppSnapShot :: SnapShot crypto -> PDoc

ppSnapShots :: SnapShots crypto -> PDoc

ppUTxO :: PrettyA (TxOut era) => UTxO era -> PDoc

ppMetadatum :: Metadatum -> PDoc

ppMetadata :: Metadata era -> PDoc

ppTx :: (PrettyA (TxBody era), PrettyA (AuxiliaryData era), PrettyA (Witnesses era)) => Tx era -> PDoc

ppBootstrapWitness :: Crypto crypto => BootstrapWitness crypto -> PDoc

ppWitnessSetHKD :: (Era era, PrettyA (Script era)) => WitnessSetHKD Identity era -> PDoc

ppSafeHash :: SafeHash crypto index -> PDoc

ppAuxiliaryDataHash :: AuxiliaryDataHash c -> PDoc

ppCompactForm :: Compactible a => (a -> PDoc) -> CompactForm a -> PDoc

ppDelegation :: Delegation c -> PDoc

ppPoolMetadata :: PoolMetadata -> PDoc

ppStakePoolRelay :: StakePoolRelay -> PDoc

ppPoolParams :: PoolParams c -> PDoc

ppWdrl :: Wdrl c -> PDoc

ppTxId :: TxId c -> PDoc

ppTxIn :: TxIn c -> PDoc

ppTxOut :: (Era era, PrettyA (Value era)) => TxOut era -> PDoc

ppDelegCert :: DelegCert c -> PDoc

ppPoolCert :: PoolCert c -> PDoc

ppGenesisDelegCert :: GenesisDelegCert c -> PDoc

ppMIRPot :: MIRPot -> PDoc

ppMIRTarget :: MIRTarget c -> PDoc

ppMIRCert :: MIRCert c -> PDoc

ppDCert :: DCert c -> PDoc

ppTxBody :: PrettyA (TxOut era) => PrettyA (PParamsDelta era) => TxBody era -> PDoc

ppWitVKey :: (Typeable kr, Crypto c) => WitVKey kr c -> PDoc

ppStakeCreds :: StakeCreds c -> PDoc

ppIPv4 :: IPv4 -> PDoc

ppIPv6 :: IPv6 -> PDoc

ppCompactAddr :: Crypto c => CompactAddr c -> PDoc

ppProtVer :: ProtVer -> PDoc

ppPParams :: PParams' Identity era -> PDoc

ppPParamsUpdate :: PParams' StrictMaybe era -> PDoc

ppUpdate :: PrettyA (PParamsDelta era) => Update era -> PDoc

ppProposedPPUpdates :: PrettyA (PParamsDelta era) => ProposedPPUpdates era -> PDoc

ppPPUpdateEnv :: PPUpdateEnv c -> PDoc

ppCoin :: Coin -> PDoc

ppDeltaCoin :: DeltaCoin -> PDoc

ppBootstrapAddress :: BootstrapAddress c -> PDoc

ppNetworkMagic :: NetworkMagic -> PDoc

ppUnparsedFields :: UnparsedFields -> PDoc

ppAddr :: Addr c -> PDoc

ppHDAddressPayload :: HDAddressPayload -> PDoc

ppRewardAcnt :: RewardAcnt c -> PDoc

ppCredential :: Credential keyrole c -> PDoc

ppPtr :: Ptr -> PDoc

ppStakeReference :: StakeReference c -> PDoc

ppGenesisCredential :: GenesisCredential c -> PDoc

ppMultiSig :: Crypto crypto => MultiSig crypto -> PDoc

ppScriptHash :: ScriptHash crypto -> PDoc

ppSlotNo :: SlotNo -> Doc ann

ppDuration :: Duration -> Doc ann

ppEpochNo :: EpochNo -> Doc ann

ppEpochSize :: EpochSize -> Doc ann

ppBlockNo :: BlockNo -> Doc ann

ppKESPeriod :: KESPeriod -> PDoc

ppOCertEnv :: OCertEnv crypto -> PDoc

ppOCert :: forall crypto. Crypto crypto => OCert crypto -> PDoc

ppOCertSignable :: forall crypto. Crypto crypto => OCertSignable crypto -> PDoc

ppHash :: Hash a b -> PDoc

ppUnitInterval :: UnitInterval -> PDoc

ppNonce :: Nonce -> PDoc

ppActiveSlotCoeff :: ActiveSlotCoeff -> PDoc

ppGlobals :: Globals -> PDoc

ppNetwork :: Network -> PDoc

ppSystemStart :: SystemStart -> PDoc

ppUrl :: Url -> PDoc

ppPort :: Port -> PDoc

ppDnsName :: DnsName -> PDoc

ppVKey :: Crypto c => VKey r c -> PDoc

ppKeyPair :: Crypto c => KeyPair r c -> PDoc

ppKeyHash :: KeyHash x c -> PDoc

ppGenDelegPair :: GenDelegPair c -> PDoc

ppGKeys :: Crypto c => GKeys c -> PDoc

ppVerKeyKES :: forall crypto. Crypto crypto => Proxy crypto -> VerKeyKES crypto -> PDoc

ppGenDelegs :: GenDelegs c -> PDoc

atWidth :: Int -> a -> (a -> PDoc) -> IO ()

Used to test pretty printing things with different widths for example: testwidth 120 ls ppLedgerState prints LedgerState, ls, with a max width of 120 columns one can use this to observe the how "pretty" a printer is at different widths

newtype Nice t

Wrap a type: t (that has a PrettyA instance) with a newtype cnstructor, so that the prettyA instance is used to Show (Nice t). Can also be used as a DerivingVia helper.

Constructors

Nice

Fields

unNice :: t

ptrace :: PrettyA t => String -> t -> a -> a

occaisionally :: Hashable a => a -> Int -> String -> String

turn on trace appromimately 1 in n times it is called.