Cardano.Ledger.Pretty

Documentation

ppString ∷ String → Doc a

ppDouble ∷ Double → Doc a

ppInteger ∷ Integer → Doc a

ppRational ∷ Rational → Doc a

ppFloat ∷ Float → Doc a

ppNatural ∷ Natural → Doc a

ppWord64 ∷ Word64 → Doc a

ppWord32 ∷ Word32 → Doc a

ppWord8 ∷ Word8 → Doc a

ppWord16 ∷ Word16 → Doc a

ppFixedPoint ∷ FixedPoint → Doc a

ppPair ∷ (t1 → PDoc) → (t2 → PDoc) → (t1, t2) → PDoc

ppSignedDSIGN ∷ Show a ⇒ a → PDoc

ppBool ∷ Bool → Doc a

ppInt ∷ Int → Doc a

isEmpty ∷ Doc ann → Bool

putDoc ∷ Doc ann → IO ()

newtype PrettyAnn

Constructors

Width Int

type Ann = [PrettyAnn]

type PDoc = Doc Ann

text ∷ Text → Doc ann

long_bech32 ∷ ByteString → Text

lazy_bech32 ∷ ByteString → Text

ppLong ∷ ByteString → PDoc

ppLazy ∷ ByteString → PDoc

equate ∷ Doc a → Doc a → Doc a

x == y

arrow ∷ (Doc a, Doc a) → Doc a

x -> y

ppSexp ∷ Text → [PDoc] → PDoc

ppSexp x [w,y,z] --> (x w y z)

ppSexp' ∷ PDoc → [PDoc] → PDoc

ppRecord ∷ Text → [(Text, PDoc)] → PDoc

ppRecord name [("a",x),("b",y),("c",z)] --> name { a = x, b = y, c = z }

ppRecord' ∷ PDoc → [(Text, PDoc)] → PDoc

puncLeft ∷ Doc ann → [Doc ann] → Doc ann → Doc ann → Doc ann

Vertical layout with commas aligned on the left hand side

ppSet ∷ (x → Doc ann) → Set x → Doc ann

ppList ∷ (x → Doc ann) → [x] → Doc ann

ppStrictSeq ∷ (a → Doc ann) → StrictSeq a → Doc ann

ppStrictMaybe ∷ (x → Doc ann) → StrictMaybe x → Doc ann

ppMaybe ∷ (x → Doc ann) → Maybe x → Doc ann

ppAssocList ∷ PDoc → (k → PDoc) → (v → PDoc) → [(k, v)] → PDoc

ppMap' ∷ PDoc → (k → PDoc) → (v → PDoc) → Map k v → PDoc

ppMap ∷ (k → PDoc) → (v → PDoc) → Map k v → PDoc

ppVMap ∷ (Vector kv k, Vector vv v) ⇒ (k → PDoc) → (v → PDoc) → VMap kv vv k v → PDoc

class PrettyA t where

Methods

prettyA ∷ t → PDoc

ppRDPair ∷ RDPair → PDoc

ppTrip ∷ Trip c → PDoc

ppUnifiedMap ∷ UMap c → PDoc

ppLastAppliedBlock ∷ LastAppliedBlock c → PDoc

ppHashHeader ∷ HashHeader c → PDoc

ppWithOrigin ∷ (t → PDoc) → WithOrigin t → PDoc

ppBHBody ∷ Crypto c ⇒ BHBody c → PDoc

ppPrevHash ∷ PrevHash c → PDoc

ppBHeader ∷ Crypto c ⇒ BHeader c → PDoc

ppBlock ∷ (PrettyA (TxSeq era), PrettyA h) ⇒ Block h era → PDoc

ppPoolDistr ∷ PoolDistr c → PDoc

ppIndividualPoolStake ∷ IndividualPoolStake c → PDoc

ppRewardUpdate ∷ RewardUpdate c → PDoc

ppRewardSnapShot ∷ RewardSnapShot c → PDoc

ppPoolRewardInfo ∷ PoolRewardInfo c → PDoc

ppFreeVars ∷ FreeVars c → PDoc

ppAns ∷ RewardAns c → PDoc

ppRewardPulser ∷ Pulser c → PDoc

ppPulsingRewUpdate ∷ PulsingRewUpdate c → PDoc

type CanPrettyPrintLedgerState era = (PrettyA (TxOut era), PrettyA (PParams era))

Constraints needed to ensure that the ledger state can be pretty printed.

ppAccountState ∷ AccountState → PDoc

ppCertState ∷ CertState era → PDoc

ppDState ∷ DState era → PDoc

ppFutureGenDeleg ∷ FutureGenDeleg c → PDoc

ppInstantaneousRewards ∷ InstantaneousRewards c → PDoc

ppShelleyPPUPState ∷ PrettyA (PParamsUpdate era) ⇒ ShelleyPPUPState era → PDoc

ppPState ∷ PState c → PDoc

ppRewardAccounts ∷ Map (Credential 'Staking c) Coin → PDoc

ppRewardType ∷ RewardType → PDoc

ppReward ∷ Reward c → PDoc

ppLeaderOnlyReward ∷ LeaderOnlyReward c → PDoc

ppIncrementalStake ∷ IncrementalStake c → PDoc

ppUTxOState ∷ (CanPrettyPrintLedgerState era, PrettyA (GovernanceState era)) ⇒ UTxOState era → PDoc

ppEpochState ∷ (CanPrettyPrintLedgerState era, PrettyA (GovernanceState era)) ⇒ EpochState era → PDoc

ppNewEpochState ∷ (CanPrettyPrintLedgerState era, PrettyA (GovernanceState era)) ⇒ NewEpochState era → PDoc

ppLedgerState ∷ (CanPrettyPrintLedgerState era, PrettyA (GovernanceState era)) ⇒ LedgerState era → PDoc

ppPerformanceEstimate ∷ PerformanceEstimate → PDoc

ppNonMyopic ∷ NonMyopic c → PDoc

ppStakeShare ∷ StakeShare → PDoc

ppHistogram ∷ Histogram → PDoc

ppLogWeight ∷ LogWeight → PDoc

ppLikelihood ∷ Likelihood → PDoc

ppStake ∷ Stake c → PDoc

ppBlocksMade ∷ BlocksMade c → PDoc

ppSnapShot ∷ SnapShot c → PDoc

ppSnapShots ∷ SnapShots c → PDoc

ppUTxO ∷ PrettyA (TxOut era) ⇒ UTxO era → PDoc

ppMetadatum ∷ Metadatum → PDoc

ppShelleyTxAuxData ∷ Era era ⇒ ShelleyTxAuxData era → PDoc

ppTx ∷ (PrettyA (TxBody era), PrettyA (TxAuxData era), PrettyA (TxWits era), EraTx era) ⇒ Tx era → PDoc

ppBootstrapWitness ∷ Crypto c ⇒ BootstrapWitness c → PDoc

ppWitnessSetHKD ∷ (Era era, PrettyA (Script era)) ⇒ ShelleyTxWits era → PDoc

ppSafeHash ∷ SafeHash c index → PDoc

ppAuxiliaryDataHash ∷ AuxiliaryDataHash c → PDoc

ppCompactForm ∷ Compactible a ⇒ (a → PDoc) → CompactForm a → PDoc

ppDelegation ∷ Delegation c → PDoc

ppPoolMetadata ∷ PoolMetadata → PDoc

ppStakePoolRelay ∷ StakePoolRelay → PDoc

ppPoolParams ∷ PoolParams c → PDoc

ppWithdrawals ∷ Withdrawals c → PDoc

ppTxId ∷ TxId c → PDoc

ppTxIn ∷ TxIn c → PDoc

ppTxOut ∷ (EraTxOut era, PrettyA (Value era)) ⇒ ShelleyTxOut era → PDoc

ppDelegCert ∷ DelegCert c → PDoc

ppPoolCert ∷ PoolCert c → PDoc

ppConstitutionalDelegCert ∷ ConstitutionalDelegCert c → PDoc

ppMIRPot ∷ MIRPot → PDoc

ppMIRTarget ∷ MIRTarget c → PDoc

ppMIRCert ∷ MIRCert c → PDoc

ppDCert ∷ DCert c → PDoc

ppTxBody ∷ (EraTxOut era, PrettyA (PParamsUpdate era), PrettyA (TxOut era)) ⇒ ShelleyTxBody era → PDoc

ppWitVKey ∷ (Typeable kr, Crypto c) ⇒ WitVKey kr c → PDoc

ppIPv4 ∷ IPv4 → PDoc

ppIPv6 ∷ IPv6 → PDoc

ppCompactAddr ∷ Crypto c ⇒ CompactAddr c → PDoc

ppVersion ∷ Version → PDoc

ppProtVer ∷ ProtVer → PDoc

ppPParams ∷ (ProtVerAtMost era 4, ProtVerAtMost era 6, EraPParams era) ⇒ PParams era → PDoc

ppPParamsUpdate ∷ (ProtVerAtMost era 4, ProtVerAtMost era 6, EraPParams era) ⇒ PParamsUpdate era → PDoc

ppUpdate ∷ PrettyA (PParamsUpdate era) ⇒ Update era → PDoc

ppProposedPPUpdates ∷ PrettyA (PParamsUpdate era) ⇒ ProposedPPUpdates era → PDoc

ppPPUpdateEnv ∷ PPUpdateEnv c → PDoc

ppCoin ∷ Coin → PDoc

ppDeltaCoin ∷ DeltaCoin → PDoc

ppBootstrapAddress ∷ BootstrapAddress c → PDoc

ppNetworkMagic ∷ NetworkMagic → PDoc

ppUnparsedFields ∷ UnparsedFields → PDoc

ppAddr ∷ Addr c → PDoc

ppHDAddressPayload ∷ HDAddressPayload → PDoc

ppRewardAcnt' ∷ RewardAcnt c → PDoc

ppCredential ∷ Credential keyrole c → PDoc

ppPtr ∷ Ptr → PDoc

ppStakeReference ∷ StakeReference c → PDoc

ppGenesisCredential ∷ GenesisCredential c → PDoc

ppMultiSig ∷ Era era ⇒ MultiSig era → PDoc

ppScriptHash ∷ ScriptHash c → PDoc

ppSlotNo ∷ SlotNo → Doc ann

ppDuration ∷ Duration → Doc ann

ppEpochNo ∷ EpochNo → Doc ann

ppEpochSize ∷ EpochSize → Doc ann

ppBlockNo ∷ BlockNo → Doc ann

ppKESPeriod ∷ KESPeriod → PDoc

ppOCertEnv ∷ OCertEnv c → PDoc

ppOCert ∷ ∀ c. Crypto c ⇒ OCert c → PDoc

ppOCertSignable ∷ ∀ c. Crypto c ⇒ OCertSignable c → PDoc

ppHash ∷ Hash a b → PDoc

ppUnitInterval ∷ UnitInterval → PDoc

ppNonce ∷ Nonce → PDoc

ppActiveSlotCoeff ∷ ActiveSlotCoeff → PDoc

ppGlobals ∷ Globals → PDoc

ppNetwork ∷ Network → PDoc

ppSystemStart ∷ SystemStart → PDoc

ppUrl ∷ Url → PDoc

ppPort ∷ Port → PDoc

ppDnsName ∷ DnsName → PDoc

ppVKey ∷ Crypto c ⇒ VKey r c → PDoc

ppKeyHash ∷ KeyHash x c → PDoc

ppGenDelegPair ∷ GenDelegPair c → PDoc

ppGKeys ∷ Crypto c ⇒ GKeys c → PDoc

ppVerKeyKES ∷ ∀ c. Crypto c ⇒ Proxy c → VerKeyKES c → PDoc

ppGenDelegs ∷ GenDelegs c → PDoc

atWidth ∷ Int → a → (a → PDoc) → IO ()

Used to test pretty printing things with different widths for example: testwidth 120 ls ppLedgerState prints LedgerState, ls, with a max width of 120 columns one can use this to observe the how "pretty" a printer is at different widths

newtype Nice t

Wrap a type: t (that has a PrettyA instance) with a newtype cnstructor, so that the prettyA instance is used to Show (Nice t). Can also be used as a DerivingVia helper.

Constructors

Nice

Fields

unNice ∷ t

ptrace ∷ PrettyA t ⇒ String → t → a → a

occaisionally ∷ Hashable a ⇒ a → Int → String → String

turn on trace appromimately 1 in n times it is called.